Remove binary image assets
